Some formating and refactoring after code climate review
