+++ storage/internal/elastic/queries.go
@@ -1,0 +1,137 @@
+package elastic
+
+// Return the query that is used to create indexes for event and log storage.
+// We only index the needed fields to optimize ES indexing performance (no dynamic mapping).
+func buildInitStorageIndexQuery() string {
+	query := `
+{
+     "settings": {
+         "refresh_interval": "1s"
+     },
+     "mappings": {
+         "logs_or_event": {
+             "_all": {"enabled": false},
+             "dynamic": "false",
+             "properties": {
+                 "clusterId": {
+                     "type": "keyword",
+                     "index": true
+                 },
+                 "deploymentId": {
+                     "type": "keyword",
+                     "index": true
+                 },
+                 "iid": {
+                     "type": "keyword",
+                     "index": true
+                 }
+             }
+         }
+     }
+}`
+	return query
+}
+
+// This ES aggregation query is built using clusterId and eventually deploymentId.
+func buildLastModifiedIndexQuery(clusterId string, deploymentId string) string {
+	var query string
+	if len(deploymentId) == 0 {
+		query = `
+{
+    "aggs" : {
+        "logs_or_events" : {
+            "filter" : {
+				"term": { "clusterId": "` + clusterId + `" }
+            },
+            "aggs" : {
+                "last_index" : { "max" : { "field" : "iid" } }
+            }
+        }
+    }
+}`
+	} else {
+		query = `
+{
+    "aggs" : {
+        "logs_or_events" : {
+            "filter" : {
+                "bool": {
+                    "must": [
+                        { "term": { "deploymentId": "` + deploymentId + `" } },
+                        { "term": { "clusterId": "` + clusterId + `" } }
+                     ]
+                }
+            },
+            "aggs" : {
+                "last_index" : { "max" : { "field" : "iid" } }
+            }
+        }
+    }
+}`
+	}
+	return query
+}
+
+// This ES range query is built using 'waitIndex' and eventually 'maxIndex' and filtered using 'clusterId' and eventually 'deploymentId'.
+func getListQuery(clusterID string, deploymentId string, waitIndex uint64, maxIndex uint64) string {
+
+	var rangeQuery, query string
+	if maxIndex > 0 {
+		rangeQuery = `
+            {
+               "range":{
+                  "iid":{
+                     "gt": "` + getSortableStringFromUint64(waitIndex) + `",
+					 "lte": "` + getSortableStringFromUint64(maxIndex) + `"
+                  }
+               }
+            }`
+	} else {
+		rangeQuery = `
+            {
+               "range":{
+                  "iid":{
+                     "gt": "` + getSortableStringFromUint64(waitIndex) + `"
+                  }
+               }
+            }`
+	}
+	if len(deploymentId) == 0 {
+		query = `
+{
+   "query":{
+      "bool":{
+         "must":[
+            {
+               "term":{
+                  "clusterId":"` + clusterID + `"
+               }
+            },` + rangeQuery + `
+         ]
+      }
+   }
+}`
+	} else {
+		query = `
+{
+   "query":{
+      "bool":{
+         "must":[
+            {
+               "term":{
+                  "clusterId":"` + clusterID + `"
+               }
+            },
+            {
+               "term":{
+                  "deploymentId":"` + deploymentId + `"
+               }
+            },` + rangeQuery + `
+         ]
+      }
+   }
+}`
+	}
+	return query
+}
+
