+++ storage/internal/elastic/store.go
@@ -33,7 +33,6 @@ import (
 	"io/ioutil"
 	"math"
 	"reflect"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -85,37 +84,6 @@ type lastIndexAggregation struct {
 }
 type stringValue struct {
 	value string `json:"value"`
-}
-
-// Precompiled regex to extract storeType and timestamp from a key of the form: "_yorc/logs/MyApp/2020-06-07T21:03:17.812178429Z".
-var storeTypeAndTimestampRegex = regexp.MustCompile(`(?m)\_yorc\/(\w+)\/.+\/(.*)`)
-
-// Parse a key of form  "_yorc/logs/MyApp/2020-06-07T21:03:17.812178429Z" to get the store type (logs|events) and the timestamp.
-func extractStoreTypeAndTimestamp(k string) (storeType string, timestamp string) {
-	res := storeTypeAndTimestampRegex.FindAllStringSubmatch(k, -1)
-	for i := range res {
-		storeType = res[i][1]
-		timestamp = res[i][2]
-	}
-	return storeType, timestamp
-}
-
-// Precompiled regex to extract storeType and deploymentId from a key of the form: "_yorc/events/" or "_yorc/logs/MyApp" or "_yorc/logs/MyApp/".
-var storeTypeAndDeploymentIDRegex = regexp.MustCompile(`(?m)\_yorc\/(\w+)\/?(.+)?\/?`)
-
-// Parse a key of form "_yorc/events/" or "_yorc/logs/MyApp" or "_yorc/logs/MyApp/" to get the store type (logs|events) and eventually the deploymentId.
-func extractStoreTypeAndDeploymentID(k string) (storeType string, deploymentID string) {
-	res := storeTypeAndDeploymentIDRegex.FindAllStringSubmatch(k, -1)
-	for i := range res {
-		storeType = res[i][1]
-		if len(res[i]) == 3 {
-			deploymentID = res[i][2]
-			if strings.HasSuffix(deploymentID, "/") {
-				deploymentID = deploymentID[:len(deploymentID)-1]
-			}
-		}
-	}
-	return storeType, deploymentID
 }
 
 // Get the tag for this field (for internal usage only: fatal if not found !).
@@ -250,107 +218,6 @@ func NewStore(cfg config.Configuration, storeConfig config.Store) store.Store {
 	return &elasticStore{encoding.JSON, esClient, clusterID, esCfg}
 }
 
-// Return the query that is used to create indexes for event and log storage.
-// We only index the needed fields to optimize ES indexing performance (no dynamic mapping).
-func buildInitStorageIndexQuery() string {
-	query := `
-{
-     "settings": {
-         "refresh_interval": "1s"
-     },
-     "mappings": {
-         "logs_or_event": {
-             "_all": {"enabled": false},
-             "dynamic": "false",
-             "properties": {
-                 "clusterId": {
-                     "type": "keyword",
-                     "index": true
-                 },
-                 "deploymentId": {
-                     "type": "keyword",
-                     "index": true
-                 },
-                 "iid": {
-                     "type": "keyword",
-                     "index": true
-                 }
-             }
-         }
-     }
-}`
-	return query
-}
-
-// Init ES index for logs or events storage: create it if not found.
-func initStorageIndex(esClient *elasticsearch6.Client, indexName string) {
-
-	log.Printf("Checking if index <%s> already exists", indexName)
-
-	// check if the sequences index exists
-	req := esapi.IndicesExistsRequest{
-		Index:           []string{indexName},
-		ExpandWildcards: "none",
-		AllowNoIndices:  &pfalse,
-	}
-	res, err := req.Do(context.Background(), esClient)
-	debugESResponse("IndicesExistsRequest:"+indexName, res, err)
-	defer res.Body.Close()
-
-	if res.StatusCode == 200 {
-		log.Printf("Indice %s was found, nothing to do !", indexName)
-	}
-
-	if res.StatusCode == 404 {
-		log.Printf("Indice %s was not found, let's create it !", indexName)
-
-		requestBodyData := buildInitStorageIndexQuery()
-
-		// indice doest not exist, let's create it
-		req := esapi.IndicesCreateRequest{
-			Index: indexName,
-			Body:  strings.NewReader(requestBodyData),
-		}
-		res, err := req.Do(context.Background(), esClient)
-		debugESResponse("IndicesCreateRequest:"+indexName, res, err)
-		defer res.Body.Close()
-		if res.IsError() {
-			var rsp map[string]interface{}
-			json.NewDecoder(res.Body).Decode(&rsp)
-			log.Printf("Response for IndicesCreateRequest (%s) : %+v", indexName, rsp)
-		}
-
-	}
-
-}
-
-// Just to display index settings at startup.
-func debugIndexSetting(esClient *elasticsearch6.Client, indexName string) {
-	log.Debugf("Get settings for index <%s>", indexName)
-	req := esapi.IndicesGetSettingsRequest{
-		Index:  []string{indexName},
-		Pretty: true,
-	}
-	res, err := req.Do(context.Background(), esClient)
-	debugESResponse("IndicesGetSettingsRequest:"+indexName, res, err)
-	defer res.Body.Close()
-}
-
-// Debug the ES response.
-func debugESResponse(msg string, res *esapi.Response, err error) {
-	if err != nil {
-		log.Debugf("[%s] Error while requesting ES : %+v", msg, err)
-	} else if res.IsError() {
-		var rsp map[string]interface{}
-		json.NewDecoder(res.Body).Decode(&rsp)
-		log.Debugf("[%s] Response Error while requesting ES (%d): %+v", msg, res.StatusCode, rsp)
-	} else {
-		var rsp map[string]interface{}
-		json.NewDecoder(res.Body).Decode(&rsp)
-		log.Debugf("[%s] Success ES response (%d): %+v", msg, res.StatusCode, rsp)
-	}
-}
-
 // Perform a refresh query on ES cluster for this particular index.
 func (s *elasticStore) refreshIndex(indexName string) {
 	req := esapi.IndicesRefreshRequest{
@@ -385,20 +252,6 @@ func (s *elasticStore) buildElasticDocument(k string, v interface{}) (string, []
 	raw = appendJSONInBytes(raw, []byte(a))
 
 	return storeType, raw, nil
-}
-
-// We need to append JSON directly into []byte to avoid useless and costly  marshaling / unmarshaling.
-func appendJSONInBytes(a []byte, v []byte) []byte {
-	last := len(a) - 1
-	lastByte := a[last]
-	// just append v at the end
-	a = append(a, v...)
-	// then slice
-	for i, s := range v {
-		a[last+i] = s
-	}
-	a[last+len(v)] = lastByte
-	return a
 }
 
 // Set index a document (log or event) into ES.
@@ -540,9 +393,9 @@ func (s *elasticStore) eventuallyAppendValueToBulkRequest(body []byte, kv store.
 
 	// The bulk action
 	index := `{"index":{"_index":"` + s.getIndexName(storeType) + `","_type":"logs_or_event"}}`
-	// 2 = len("\n\n")
-	// FIXME: not able to make it work defining the slice length
+	// TODO: not able to make it work defining the slice length
 	// 2020/06/09 02:54:20 [FATAL] failed to migrate data from Consul for root path:"_yorc/logs" in store with name:"elastic": Error while sending bulk request, response code was <500> and response message was <[500 Internal Server Error] {"error":{"root_cause":[{"type":"char_conversion_exception","reason":"Invalid UTF-32 character 0x7b21696e (above 0x0010ffff) at char #84, byte #339)"}],"type":"char_conversion_exception","reason":"Invalid UTF-32 character 0x7b21696e (above 0x0010ffff) at char #84, byte #339)"},"status":500}>
+	// 2 = len("\n\n")
 	//bulkOperation := make([]byte, len(index) + len(document) + 2)
 	bulkOperation := make([]byte, 0)
 	bulkOperation = append(bulkOperation, index...)
@@ -625,15 +478,6 @@ func (s *elasticStore) GetLastModifyIndex(k string) (uint64, error) {
 
 	return s.internalGetLastModifyIndex(s.getIndexName(indexName), deploymentID)
 
-}
-
-func parseInt64StringToUint64(value string) (uint64, error) {
-	valueInt, err := strconv.ParseInt(value, 10, 64)
-	if err != nil {
-		return 0, errors.Wrapf(err, "Not able to parse string: %s to int64, error was: %+v", value, err)
-	}
-	result := uint64(valueInt)
-	return result, nil
 }
 
 // The last index is found by querying ES using aggregation and a 0 size request.
@@ -753,120 +597,6 @@ func (s *elasticStore) List(ctx context.Context, k string, waitIndex uint64, tim
 	}
 	log.Debugf("List called result k: %s, waitIndex: %d, timeout: %v, LastIndex: %d, len(values): %d", k, waitIndex, timeout, lastIndex, len(values))
 	return values, lastIndex, err
-}
-
-// This ES aggregation query is built using clusterId and eventually deploymentId.
-func buildLastModifiedIndexQuery(clusterId string, deploymentId string) string {
-	var query string
-	if len(deploymentId) == 0 {
-		query = `
-{
-    "aggs" : {
-        "logs_or_events" : {
-            "filter" : {
-				"term": { "clusterId": "` + clusterId + `" }
-            },
-            "aggs" : {
-                "last_index" : { "max" : { "field" : "iid" } }
-            }
-        }
-    }
-}`
-	} else {
-		query = `
-{
-    "aggs" : {
-        "logs_or_events" : {
-            "filter" : {
-                "bool": {
-                    "must": [
-                        { "term": { "deploymentId": "` + deploymentId + `" } },
-                        { "term": { "clusterId": "` + clusterId + `" } }
-                     ]
-                }
-            },
-            "aggs" : {
-                "last_index" : { "max" : { "field" : "iid" } }
-            }
-        }
-    }
-}`
-	}
-	return query
-}
-
-// The max uint is 9223372036854775807 (19 cars), this is the maximum nano for a time (2262-04-12 00:47:16.854775807 +0100 CET).
-// Since we use nanotimestamp as ID for events and logs, and since we store this ID as string in ES index, we must ensure that
-// the string will be comparable.
-func getSortableStringFromUint64(nanoTimestamp uint64) string {
-	nanoTimestampStr := strconv.FormatUint(nanoTimestamp, 10)
-	if len(nanoTimestampStr) < 19 {
-		nanoTimestampStr = strings.Repeat("0", 19-len(nanoTimestampStr)) + nanoTimestampStr
-	}
-	return nanoTimestampStr
-}
-
-// This ES range query is built using 'waitIndex' and eventually 'maxIndex' and filtered using 'clusterId' and eventually 'deploymentId'.
-func getListQuery(clusterID string, deploymentId string, waitIndex uint64, maxIndex uint64) string {
-
-	var rangeQuery, query string
-	if maxIndex > 0 {
-		rangeQuery = `
-            {
-               "range":{
-                  "iid":{
-                     "gt": "` + getSortableStringFromUint64(waitIndex) + `",
-					 "lte": "` + getSortableStringFromUint64(maxIndex) + `"
-                  }
-               }
-            }`
-	} else {
-		rangeQuery = `
-            {
-               "range":{
-                  "iid":{
-                     "gt": "` + getSortableStringFromUint64(waitIndex) + `"
-                  }
-               }
-            }`
-	}
-	if len(deploymentId) == 0 {
-		query = `
-{
-   "query":{
-      "bool":{
-         "must":[
-            {
-               "term":{
-                  "clusterId":"` + clusterID + `"
-               }
-            },` + rangeQuery + `
-         ]
-      }
-   }
-}`
-	} else {
-		query = `
-{
-   "query":{
-      "bool":{
-         "must":[
-            {
-               "term":{
-                  "clusterId":"` + clusterID + `"
-               }
-            },
-            {
-               "term":{
-                  "deploymentId":"` + deploymentId + `"
-               }
-            },` + rangeQuery + `
-         ]
-      }
-   }
-}`
-	}
-	return query
 }
 
 // Query ES for events or logs specifying the expected results 'size' and the sort 'order'.
